arch: Provide stubbed host-build configuration.
The sample environ is set up for WSL2 debian and required
installing the host-side newlib. This still needs quite
a bit of cleaning up, but wanted to toss it up for visibility
(and in case I just drop it/forget).

I figure this could act as:
1) Documentation for the arch interface
2) An alternative build test for kos common code

From building this I was able to see some of our inappropriate
newlib integration, and a LOT of warnings are generated that
show where we may be unintentionally using 32bit fixed values
for platform-dependent sizes.